requirements.txt
asgiref==3.7.2
dj-rest-auth==4.0.1
Django==4.2.3
django-cors-headers==4.2.0
django-filter==23.2
django-utils-six==2.0
djangorestframework==3.14.0
Markdown==3.4.4
psycopg2-binary==2.9.6
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
